chore: move some auth middleware stuff to shareable base class
